requirements/ci.txt
-r base.txt

coverage==4.4
django-debug-toolbar==1.8
django-extensions
